started on reply functionality still in progress

+++ src/components/CommentCard/Comment.js
@@ -4,18 +4,39 @@ import { orange } from "@mui/material/colors";
 import { NavLink } from "react-router-dom";
 import { useContext } from "react";
 import { AuthContext } from "../../contexts/AuthContext";
-import { IconButton } from "@mui/material";
+import { IconButton, Button } from "@mui/material";
 import { Delete } from "@mui/icons-material";
 import * as phoneService from "../../services/phoneService";
+import SendIcon from "@mui/icons-material/Send";
 
-const Comment = ({ commentText, createdAt, commentId, owner, phoneId }) => {
+const Comment = ({
+  commentText,
+  createdAt,
+  commentId,
+  owner,
+  phoneId,
+  setCommentState,
+  replies,
+}) => {
   const { user } = useContext(AuthContext);
   const commentOwner = Boolean(user._id === owner._id);
 
   const removeCommentHandler = (e) => {
     e.preventDefault();
 
-    phoneService.destroyComment({ phoneId, commentId });
+    const confirmation = window.confirm(
+      "Are you sure you want to remove this comment?"
+    );
+
+    if (confirmation) {
+      phoneService.destroyComment({ phoneId, commentId });
+
+      setCommentState(true);
+    }
+  };
+
+  const replyCommentHandler = (e) => {
+    e.preventDefault();
   };
 
   const removeBtn = (
@@ -26,25 +47,74 @@ const Comment = ({ commentText, createdAt, commentId, owner, phoneId }) => {
     </div>
   );
 
+  const replyDiv = (
+    <div className="replyInfo">
+      <div className="add-comment">
+        <textarea
+          className="comment-section"
+          name="comment"
+          placeholder="comment..."
+          id=""
+          cols="30"
+          rows="10"
+        ></textarea>
+        <Button
+          onClick={replyCommentHandler}
+          variant="contained"
+          endIcon={<SendIcon />}
+        >
+          replys
+        </Button>
+      </div>
+    </div>
+  );
+
   return (
-    <article className="comment">
-      <div className="comment-display">
-        <NavLink to={`/profiles/${owner.username}`} className={"avatar-link"}>
-          <Avatar sx={{ bgcolor: orange[500] }}>
-            {owner.username[0].toUpperCase()}
-          </Avatar>
-        </NavLink>
-      </div>
+    <>
+      <article className="comment">
+        <div className="comment-display">
+          <NavLink to={`/profiles/${owner.username}`} className={"avatar-link"}>
+            <Avatar sx={{ bgcolor: orange[500] }}>
+              {owner.username[0].toUpperCase()}
+            </Avatar>
+          </NavLink>
+        </div>
 
-      <div className="comment-dipslay-info">
-        <h3>
-          {owner.username} <span className="comment-date">{createdAt}</span>
-        </h3>
-        <p>{commentText}</p>
-      </div>
+        <div className="comment-dipslay-info">
+          <h3>
+            {owner.username} <span className="comment-date">{createdAt}</span>
+          </h3>
+          <p>{commentText}</p>
+        </div>
 
-      {commentOwner && removeBtn}
-    </article>
+        {commentOwner && removeBtn}
+
+        <Button variant="text">Reply</Button>
+
+        <div className="replyDiv">
+          <textarea
+            className="reply-section"
+            name="reply"
+            placeholder="reply..."
+            id=""
+            cols="30"
+            rows="10"
+          ></textarea>
+          <Button
+            className="reply-btn"
+            onClick={replyCommentHandler}
+            variant="contained"
+            endIcon={<SendIcon />}
+          >
+            Reply
+          </Button>
+        </div>
+      </article>
+
+      <article className="replies">
+        {replies?.length > 0 ? replyDiv : `Replies: ${replies?.length}`}
+      </article>
+    </>
   );
 };
